refactor(particles): merge float/u32 particle state chunks
The float and u32 readInput/writeOutput chunks differed only in the texel
cast, so fold them into single particleInput / particleOutput chunks. The
GLSL chunk handles the difference internally with a small #ifdef
CAPS_TEXTURE_FLOAT_RENDERABLE around just the cast; the WGSL chunk is the
float logic (WebGPU is always float-renderable). Include sites in both
backends are now identical and the per-format branch is gone from them.
