--- src/scene/shader-lib/glsl/chunks/particle/frag/particleOutputFloat.js
@@ -1,9 +1,0 @@
-export default /* glsl */`
-void writeOutput() {
-    if (gl_FragCoord.y<1.0) {
-        gl_FragColor = vec4(outPos, (outAngle + 1000.0) * visMode);
-    } else {
-        gl_FragColor = vec4(outVel, outLife);
-    }
-}
-`;
