+++ src/scene/shader-lib/glsl/collections/particle-chunks-glsl.js
@@ -3,10 +3,8 @@ import particleVS from '../chunks/particle/vert/particle.js';
 import particleAnimFrameClampVS from '../chunks/particle/vert/particleAnimFrameClamp.js';
 import particleAnimFrameLoopVS from '../chunks/particle/vert/particleAnimFrameLoop.js';
 import particleAnimTexVS from '../chunks/particle/vert/particleAnimTex.js';
-import particleInputFloatPS from '../chunks/particle/frag/particleInputFloat.js';
-import particleInputU32PS from '../chunks/particle/frag/particleInputU32.js';
-import particleOutputFloatPS from '../chunks/particle/frag/particleOutputFloat.js';
-import particleOutputU32PS from '../chunks/particle/frag/particleOutputU32.js';
+import particleInputPS from '../chunks/particle/frag/particleInput.js';
+import particleOutputPS from '../chunks/particle/frag/particleOutput.js';
 import particleUpdaterAABBPS from '../chunks/particle/frag/particleUpdaterAABB.js';
 import particleUpdaterEndPS from '../chunks/particle/frag/particleUpdaterEnd.js';
 import particleUpdaterInitPS from '../chunks/particle/frag/particleUpdaterInit.js';
@@ -48,10 +46,8 @@ export const particleChunksGLSL = {
     particleAnimFrameClampVS,
     particleAnimFrameLoopVS,
     particleAnimTexVS,
-    particleInputFloatPS,
-    particleInputU32PS,
-    particleOutputFloatPS,
-    particleOutputU32PS,
+    particleInputPS,
+    particleOutputPS,
     particleUpdaterAABBPS,
     particleUpdaterEndPS,
     particleUpdaterInitPS,
