+++ src/scene/shader-lib/wgsl/chunks/particle/frag/particleInput.js
