+++ src/scene/shader-lib/wgsl/chunks/particle/frag/particleOutput.js
