+++ src/scene/shader-lib/wgsl/collections/particle-chunks-wgsl.js
@@ -3,8 +3,8 @@ import particleVS from '../chunks/particle/vert/particle.js';
 import particleAnimFrameClampVS from '../chunks/particle/vert/particleAnimFrameClamp.js';
 import particleAnimFrameLoopVS from '../chunks/particle/vert/particleAnimFrameLoop.js';
 import particleAnimTexVS from '../chunks/particle/vert/particleAnimTex.js';
-import particleInputFloatPS from '../chunks/particle/frag/particleInputFloat.js';
-import particleOutputFloatPS from '../chunks/particle/frag/particleOutputFloat.js';
+import particleInputPS from '../chunks/particle/frag/particleInput.js';
+import particleOutputPS from '../chunks/particle/frag/particleOutput.js';
 import particleUpdaterAABBPS from '../chunks/particle/frag/particleUpdaterAABB.js';
 import particleUpdaterEndPS from '../chunks/particle/frag/particleUpdaterEnd.js';
 import particleUpdaterInitPS from '../chunks/particle/frag/particleUpdaterInit.js';
@@ -46,8 +46,8 @@ export const particleChunksWGSL = {
     particleAnimFrameClampVS,
     particleAnimFrameLoopVS,
     particleAnimTexVS,
-    particleInputFloatPS,
-    particleOutputFloatPS,
+    particleInputPS,
+    particleOutputPS,
     particleUpdaterAABBPS,
     particleUpdaterEndPS,
     particleUpdaterInitPS,
